+++ backend/api/routes/process_tree/discoverTree.py
@@ -1,6 +1,7 @@
 import multiprocessing
 from typing import Any, List
 
+from cortado_core.process_tree_utils.reduction import apply_reduction_rules
 from cortado_core.utils.sequentializations import generate_sequentializations
 from cortado_core.utils.split_graph import Group
 
@@ -26,12 +27,6 @@ class InputDiscoverProcessModelFromVariants(BaseModel):
     variants: List[Any]
 
 
-@router.post("/discoverProcessModelFromVariants")
-async def discover_process_model(d: InputDiscoverProcessModelFromVariants):
-    variants = [v["value"]["events"] for v in d.variants]
-    return discover_process_model_from_variants(variants)
-
-
 def discover_process_model_from_variants(variants):
     log = EventLog()
     for v in variants:
@@ -43,6 +38,7 @@ def discover_process_model_from_variants(variants):
             t.append(event)
         log.append(t)
     pt: ProcessTree = discover_process_tree_inductive(log)
+    apply_reduction_rules(pt)
     res = process_tree_to_dict(pt)
     return res
 
@@ -71,14 +67,6 @@ class InputAddVariantsToProcessModel(BaseModel):
     fitting_variants: List[Any]
     variants_to_add: List[Any]
     pt: dict
-
-
-# TODO this endpoint is currently unused, we have to decide if we want to delete it
-@router.post("/addVariantsToProcessModel")
-async def add_simple_variants_to_process_model(d: InputAddVariantsToProcessModel):
-    fitting_variants = [v["events"] for v in d.fitting_variants]
-    to_add = [v["events"] for v in d.variants_to_add]
-    return add_variants_to_process_model(d.pt, fitting_variants, to_add, PoolFactory.instance().get_pool())
 
 
 @router.post("/addConcurrencyVariantsToProcessModel")